+++ modules/apps/commerce/commerce-shipping-web/src/main/java/com/liferay/commerce/shipping/web/internal/servlet/taglib/ui/CommerceShippingMethodDetailsScreenNavigationEntry.java
@@ -15,6 +15,7 @@
 package com.liferay.commerce.shipping.web.internal.servlet.taglib.ui;
 
 import com.liferay.commerce.model.CommerceShippingMethod;
+import com.liferay.commerce.shipping.web.internal.servlet.taglib.ui.constants.CommerceShippingScreenNavigationConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
