+++ modules/apps/commerce/commerce-shipping-web/src/main/resources/META-INF/resources/init.jsp
@@ -29,7 +29,7 @@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 page import="com.liferay.commerce.model.CommerceShippingMethod" %><%@
 page import="com.liferay.commerce.shipping.web.internal.display.context.CommerceShippingMethodsDisplayContext" %><%@
 page import="com.liferay.commerce.shipping.web.internal.frontend.CommerceShippingRestrictionsPageClayTable" %><%@
-page import="com.liferay.commerce.shipping.web.internal.servlet.taglib.ui.CommerceShippingScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.shipping.web.internal.servlet.taglib.ui.constants.CommerceShippingScreenNavigationConstants" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
 page import="com.liferay.portal.kernel.bean.BeanParamUtil" %><%@
 page import="com.liferay.portal.kernel.util.Constants" %><%@
