+++ modules/apps/commerce/commerce-subscription-web/src/main/resources/META-INF/resources/init.jsp
@@ -41,8 +41,8 @@ page import="com.liferay.commerce.subscription.web.internal.display.context.CPDe
 page import="com.liferay.commerce.subscription.web.internal.display.context.CPInstanceSubscriptionInfoDisplayContext" %><%@
 page import="com.liferay.commerce.subscription.web.internal.display.context.CommerceSubscriptionContentDisplayContext" %><%@
 page import="com.liferay.commerce.subscription.web.internal.display.context.CommerceSubscriptionEntryDisplayContext" %><%@
-page import="com.liferay.commerce.subscription.web.internal.frontend.CommerceSubscriptionDataSetConstants" %><%@
-page import="com.liferay.commerce.subscription.web.internal.servlet.taglib.ui.CommerceSubscriptionEntryScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.subscription.web.internal.frontend.constants.CommerceSubscriptionDataSetConstants" %><%@
+page import="com.liferay.commerce.subscription.web.internal.servlet.taglib.ui.constants.CommerceSubscriptionEntryScreenNavigationConstants" %><%@
 page import="com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem" %><%@
 page import="com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemList" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
