+++ modules/apps/commerce/commerce-tax-engine-fixed-web/src/main/java/com/liferay/commerce/tax/engine/fixed/web/internal/frontend/CommerceTaxRateSettingClayTableDataSetDisplayView.java
@@ -14,6 +14,7 @@
 
 package com.liferay.commerce.tax.engine.fixed.web.internal.frontend;
 
+import com.liferay.commerce.tax.engine.fixed.web.internal.frontend.constants.CommerceTaxRateSettingDataSetConstants;
 import com.liferay.frontend.taglib.clay.data.set.ClayDataSetDisplayView;
 import com.liferay.frontend.taglib.clay.data.set.view.table.BaseTableClayDataSetDisplayView;
 import com.liferay.frontend.taglib.clay.data.set.view.table.ClayTableSchema;
