+++ modules/apps/commerce/commerce-tax-engine-fixed-web/src/main/resources/META-INF/resources/init.jsp
@@ -32,7 +32,7 @@ page import="com.liferay.commerce.tax.engine.fixed.model.CommerceTaxFixedRateAdd
 page import="com.liferay.commerce.tax.engine.fixed.web.internal.display.context.CommerceTaxFixedRateAddressRelsDisplayContext" %><%@
 page import="com.liferay.commerce.tax.engine.fixed.web.internal.display.context.CommerceTaxFixedRatesDisplayContext" %><%@
 page import="com.liferay.commerce.tax.engine.fixed.web.internal.frontend.CommerceTaxRateClayTable" %><%@
-page import="com.liferay.commerce.tax.engine.fixed.web.internal.frontend.CommerceTaxRateSettingDataSetConstants" %><%@
+page import="com.liferay.commerce.tax.engine.fixed.web.internal.frontend.constants.CommerceTaxRateSettingDataSetConstants" %><%@
 page import="com.liferay.commerce.tax.model.CommerceTaxMethod" %><%@
 page import="com.liferay.petra.string.StringPool" %><%@
 page import="com.liferay.portal.kernel.language.LanguageUtil" %><%@
