+++ modules/apps/commerce/commerce-warehouse-web/src/main/java/com/liferay/commerce/warehouse/web/internal/servlet/taglib/ui/CommerceInventoryWarehouseGeneralFormNavigatorCategory.java
@@ -14,6 +14,7 @@
 
 package com.liferay.commerce.warehouse.web.internal.servlet.taglib.ui;
 
+import com.liferay.commerce.warehouse.web.internal.servlet.taglib.ui.constants.CommerceInventoryWarehouseFormNavigatorConstants;
 import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
 import com.liferay.portal.kernel.language.LanguageUtil;
